Added option for element-specific debug output
git-svn-id: http://svn.sintef.no/trondheim/IFEM/trunk@989 e10b68d5-8a6e-419e-a041-bce267b0401d

+++ src/ASM/SAMpatchPara.C
@@ -196,7 +196,7 @@ bool SAMpatchPara::getElmEqns (IntVec& meen, int iel, int nedof) const
     for (int j = madof[node-1]; j < madof[node]; j++)
       meen.push_back(j);
   }
-  if (meen.size() == nedof) return true;
+  if ((int)meen.size() == nedof) return true;
 
   std::cerr <<"SAMpatchPara::getElmEqns: Invalid element matrix dimension "
 	    << nedof <<" (should have been "<< meen.size() <<")"<< std::endl;
